Working with integrating the new camera class
