+++ rov/rov.py
@@ -18,6 +18,8 @@ from movement import controller
 
 from sensors import Pressure, IMU
 
+from camera import Cameras
+
 
 class ROV(object):
 
@@ -43,7 +45,7 @@ class ROV(object):
         self.cameras = Cameras(
             resolution='640x480',
             framerate=30,
-            port_start=8080,
+            port=8080,
             brightness=16,
             contrast=32
         )
